Add declaration codegen + changes for declarations
Since we had many types of values stored in symbol table not all of
which require (or can work with) the load instruction, make a
create_load function which does the checking.

Change cond_expr_ops codegen since it was directly using load to pushing
a scope for the result

Add push_scope and pop_scope member functions to func_scope

Every function now first generates an allocs block which is used for
storing all the alloca's it generates. This is done because allocas
were getting added to the end of the entry block when a declaration was
present after the entry block was terminated. This allocs block
unconditionally branched to the entry block and the builder is set to
before this branch instruction