Make scope w/ RAII, fix empty block deletions, bug fix
new_scope() method returns an auto_scope which pushes scope in the
constructor and pops it in the destructor. push_scope and pop_scope are
now private. All uses of them have been replaced with proper new_scope
use

Earlier we would delete basic blocks if they were empty, but that causes
problems if there is any branch to that block. Use the
replaceAllUsesWith function of the Value class to amend this.

Earlier the return_stmt's codegen wouldn't add a new empty block after
adding the ret instruction, thus violating the (implicit) invariant
codegen is maintaining. This wasn't an issue earlier because the
statements after it wouldn't be executed anyway. But with the addition
of goto, it's possible to bypass the return statement and then codegen
has problems